Refactor: BackupHistory 서비스, enum 구조 변경 [#43]
