+++ src/main/java/com/hrbank3/hrbank3/entity/enums/BackupStatus.java
@@ -1,4 +1,4 @@
-package com.hrbank3.hrbank3.entity;
+package com.hrbank3.hrbank3.entity.enums;
 
 public enum BackupStatus {
   IN_PROGRESS, // 진행중
